Renaming packages for check version service repository
